#70: Moved theme components to own theme module
