+++ gui/src/app/pages/pages-routing.module.ts
@@ -17,7 +17,7 @@ import {AddItemExternallyComponent} from './add-item-externally/add-item-externa
 import {AdminSettingsComponent} from './admin-settings/admin-settings.component';
 import {AdminUsersComponent} from './admin-users/admin-users.component';
 import {GlobalSearchComponent} from "./global-search/global-search.component";
-import {AppMainComponent} from "./app.main.component";
+import {AppMainComponent} from "../theme/app-main/app-main.component";
 
 const routes: Routes = [{
   path: '',
