+++ gui/src/app/theme/app-footer/app-footer.component.html
@@ -1,5 +1,5 @@
 <div class="layout-footer">
-    <img src="assets/icons/icon-192x192.png" alt="Logo" height="20" class="mr-2"/>
+    <img src="../../../assets/icons/icon-192x192.png" alt="Logo" height="20" class="mr-2"/>
     by
     <span class="font-medium ml-2">PrimeNG</span>
 </div>
