+++ gui/src/app/theme/app-footer/app-footer.component.ts
@@ -1,9 +1,9 @@
 import { Component } from '@angular/core';
-import { AppMainComponent } from './app.main.component';
+import { AppMainComponent } from '../app-main/app-main.component';
 
 @Component({
     selector: 'app-footer',
-    templateUrl: './app.footer.component.html'
+    templateUrl: './app-footer.component.html'
 })
 export class AppFooterComponent{
     constructor(public appMain: AppMainComponent) {}
