+++ gui/src/app/theme/app-menu/app-menu.component.ts
@@ -1,9 +1,9 @@
 import {Component, OnInit} from '@angular/core';
-import {AppMainComponent} from './app.main.component';
+import {AppMainComponent} from '../app-main/app-main.component';
 
 @Component({
   selector: 'app-menu',
-  templateUrl: './app.menu.component.html',
+  templateUrl: './app-menu.component.html',
 })
 export class AppMenuComponent implements OnInit {
