+++ gui/src/app/theme/app-topbar/app-topbar.component.html
@@ -1,6 +1,6 @@
 <div class="layout-topbar">
     <a class="layout-topbar-logo" routerLink="">
-        <img src="assets/icons/icon-192x192.png" alt="logo">
+        <img src="../../../assets/icons/icon-192x192.png" alt="logo">
         <span>Collection manager</span>
     </a>
