+++ gui/src/app/theme/app-topbar/app-topbar.component.ts
@@ -1,13 +1,13 @@
 import { Component, OnDestroy } from '@angular/core';
-import { AppMainComponent } from './app.main.component';
+import { AppMainComponent } from '../app-main/app-main.component';
 import { Subscription } from 'rxjs';
 import { MenuItem } from 'primeng/api';
 
 @Component({
     selector: 'app-topbar',
-    templateUrl: './app.topbar.component.html'
+    templateUrl: './app-topbar.component.html'
 })
-export class AppTopBarComponent {
+export class AppTopbarComponent {
 
     items: MenuItem[];
