+++ gui/src/app/theme/theme.module.ts
@@ -1,0 +1,35 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import {AppFooterComponent} from "./app-footer/app-footer.component";
+import {AppConfigComponent} from "./app-config/app-config.component";
+import {AppMainComponent} from "./app-main/app-main.component";
+import {AppMenuitemComponent} from "./app-menuitem/app-menuitem.component";
+import {AppTopbarComponent} from "./app-topbar/app-topbar.component";
+import {RadioButtonModule} from "primeng/radiobutton";
+import {RouterModule} from "@angular/router";
+import {InputSwitchModule} from "primeng/inputswitch";
+import {AppMenuComponent} from "./app-menu/app-menu.component";
+import {FormsModule} from "@angular/forms";
+
+
+
+@NgModule({
+  declarations: [
+    AppFooterComponent,
+    AppConfigComponent,
+    AppMainComponent,
+    AppMenuitemComponent,
+    AppTopbarComponent,
+    AppMenuComponent],
+  exports: [
+
+  ],
+  imports: [
+    CommonModule,
+    RadioButtonModule,
+    RouterModule,
+    InputSwitchModule,
+    FormsModule
+  ]
+})
+export class ThemeModule { }
